refactor: move test hooks to utils
