+++ packages/react/test/utils.ts
@@ -12,6 +12,11 @@ import {
 import { getProvider, getSigners } from '../../core/test/utils'
 import { renderHook } from '.'
 import { CreateClientConfig, createClient } from '../src'
+import {
+  UseAccountConfig,
+  useAccount as useAccount_,
+} from '../src/hooks/accounts/useAccount'
+import { useNetwork as useNetwork_ } from '../src/hooks/accounts/useNetwork'
 
 type Config = Partial<CreateClientConfig>
 
@@ -111,4 +116,14 @@ export async function getUnclaimedTokenId(
     attempts += 1
   }
   return false
+}
+
+export function useAccount(config: UseAccountConfig = {}) {
+  const { ...values } = useAccount_(config)
+  return values
+}
+
+export function useNetwork() {
+  const { ...values } = useNetwork_()
+  return values
 }
